Add length_enabled
Length_enabled and enabled are two different things.  The length_enabled flag is
controlled by the register, and is cleared when the length counter becomes zero.

The enabled flag can be cleared by the length and volume, and is set on trigger.

This fixes sounds being cut too quickly in Tetris.